+++ src/input_rtlsdr.c
@@ -400,7 +400,7 @@ static void rtlsdr_input_get_summary_info(const ModuleContext* ctx, InputSummary
 
     add_summary_item(info, "Input Source", "%s", source_name_buf);
     add_summary_item(info, "Input Format", "8-bit Unsigned Complex (cu8)");
-    add_summary_item(info, "Input Sample Rate", "%d Hz", app->module.source_info.sample_rate);
+    add_summary_item(info, "Input Sample Rate", "%.15g Hz", (double)app->module.source_info.sample_rate);
     if (s_rtlsdr_config.gain_provided) {
         add_summary_item(info, "Gain", "%.1f dB (Manual)", (float)s_rtlsdr_config.gain / 10.0f);
     } else {
